refactor(pickers): ♻️  update utils & add date utils
