+++ src/timepicker/TimePickerTrigger.ts
@@ -24,7 +24,7 @@ export const useTimePickerTrigger = createHook<
   compose: usePickerBaseTrigger,
   keys: TIME_PICKER_TRIGGER_KEYS,
 
-  useProps(options, htmlProps) {
+  useProps(_, htmlProps) {
     return {
       "aria-haspopup": "listbox",
       ...htmlProps,
