+++ src/utils/index.ts
@@ -27,3 +27,5 @@ export function isTouch() {
       window.navigator.msMaxTouchPoints > 0,
   );
 }
+
+export * from "./date";
